+++ content-social_menu.php
@@ -2,7 +2,7 @@
 /**
  * The template that supplies our WordPress Social Menu
  * @since WP-Forge 5.5.1.7
- * @version 6.2.1.3
+ * @version 6.2.2
  */
 ?>
 <?php if ( has_nav_menu( 'social' ) ) : ?>